Corrigindo o valor de retorno da função getValor da classe Transacao

+++ SGBJava/src/main/java/sgb/model/dto/Transacao.java
@@ -20,7 +20,7 @@ public class Transacao {
         return matricula;
     }
 
-    public long getValor() {
+    public int getValor() {
         return valor;
     }
